maxsymlinklen must be signed comparison
SCCS-vsn: sys/ufs/ffs/ffs_vnops.c 7.92

+++ usr/src/sys/ufs/ffs/ffs_vnops.c
@@ -4,7 +4,7 @@
  *
  * %sccs.include.redist.c%
  *
- *	@(#)ffs_vnops.c	7.91 (Berkeley) %G%
+ *	@(#)ffs_vnops.c	7.92 (Berkeley) %G%
  */
 
 #include <sys/param.h>
@@ -208,7 +208,7 @@ ffs_read(ap)
 	type = ip->i_mode & IFMT;
 	if (type != IFDIR && type != IFREG && type != IFLNK)
 		panic("ffs_read type");
-	if (type == IFLNK && ip->i_size < vp->v_mount->mnt_maxsymlinklen)
+	if (type == IFLNK && (int)ip->i_size < vp->v_mount->mnt_maxsymlinklen)
 		panic("read short symlink");
 #endif
 	if (uio->uio_resid == 0)
